I did not make this client.
So I can not 100% guarantee that this client is safe! Download at your own risk. I have ran the client a couple of times and did not have my private clients leaked. As well this is not a crack/leak of the client as it is a public release. This client is still a wip, some modules do not work I will probably post newer updates of the client as they are released.